+++ server/routes/entries.js
@@ -3,6 +3,6 @@ import EntryController from '../controllers/EntryController';
 import Validator from '../utils/validation';
 
 const router = express.Router();
-router.post('/entries', Validator.addEntry, EntryController.addEntry);
+router.post('/', Validator.addEntry, EntryController.addEntry);
 
 export default router;
